requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
torchaudio>=2.0.0
numpy>=1.23.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
scikit-learn>=1.2.0
tensorboard>=2.13.0
streamlit>=1.24.0  # For the demo
librosa>=0.10.0    # For audio processing
pillow>=9.5.0      # For image processing
omegaconf>=2.3.0   # For configuration management
einops>=0.6.0      # For tensor operations
tabulate
tensorboardX
